Completely rewritten the README
The README has been completely rewritten to better reflect what LexiconDM is about, how to play it, compile it, file bug reports/mapset requests, how to reach the team, join The Mythotic Tinker-Station Discord, and so on. It sounds and reads much better, and flows better than the original README did, at least I think so!